Add some configure commands to some projects
Using configure options it is possible to minimize the
dependencies of a project. In the future we might add
some configure comments to some projects to make it
easier to start analyzin a project.

Note that disabling features and optional dependencies
also result in reduced coverage during the analysis.
So these are only to make it easier to get started with
adding a project to a test set.